requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=2_gnu
agenta=0.50.6=pypi_0
aiohappyeyeballs=2.6.1=pypi_0
aiohttp=3.12.15=pypi_0
aiosignal=1.4.0=pypi_0
alembic=1.16.4=pypi_0
annotated-types=0.7.0=pypi_0
anyio=4.10.0=pypi_0
appdirs=1.4.4=pypi_0
arxiv=2.2.0=pypi_0
asgiref=3.9.1=pypi_0
asttokens=3.0.0=pypi_0
async-timeout=4.0.3=pypi_0
attrs=25.3.0=pypi_0
backoff=2.2.1=pypi_0
bcrypt=4.3.0=pypi_0
beautifulsoup4=4.13.4=pypi_0
blinker=1.9.0=pypi_0
build=1.3.0=pypi_0
bzip2=1.0.8=h4bc722e_7
ca-certificates=2025.8.3=hbd8a1cb_0
cachetools=5.5.2=pypi_0
certifi=2025.8.3=pypi_0
cffi=1.17.1=pypi_0
charset-normalizer=3.4.3=pypi_0
chroma-hnswlib=0.7.6=pypi_0
chromadb=0.5.23=pypi_0
click=8.2.1=pypi_0
cohere=5.17.0=pypi_0
coloredlogs=15.0.1=pypi_0
contourpy=1.3.2=pypi_0
crewai=0.157.0=pypi_0
crewai-tools=0.55.0=pypi_0
cryptography=45.0.6=pypi_0
cycler=0.12.1=pypi_0
dataclasses-json=0.6.7=pypi_0
decorator=5.2.1=pypi_0
deprecation=2.1.0=pypi_0
diskcache=5.6.3=pypi_0
distro=1.9.0=pypi_0
docker=7.1.0=pypi_0
docstring-parser=0.17.0=pypi_0
durationpy=0.10=pypi_0
embedchain=0.1.128=pypi_0
et-xmlfile=2.0.0=pypi_0
exceptiongroup=1.3.0=pypi_0
executing=2.2.0=pypi_0
fastapi=0.116.1=pypi_0
fastavro=1.12.0=pypi_0
feedparser=6.0.12=pypi_0
filelock=3.18.0=pypi_0
flask=3.1.2=pypi_0
flatbuffers=25.2.10=pypi_0
fonttools=4.59.1=pypi_0
frozenlist=1.7.0=pypi_0
fsspec=2025.7.0=pypi_0
google-auth=2.40.3=pypi_0
googleapis-common-protos=1.70.0=pypi_0
gptcache=0.1.44=pypi_0
greenlet=3.2.4=pypi_0
grpcio=1.74.0=pypi_0
h11=0.16.0=pypi_0
h2=4.2.0=pypi_0
hf-xet=1.1.7=pypi_0
hpack=4.1.0=pypi_0
httpcore=1.0.9=pypi_0
httptools=0.6.4=pypi_0
httpx=0.28.1=pypi_0
httpx-sse=0.4.0=pypi_0
huggingface-hub=0.30.2=pypi_0
humanfriendly=10.0=pypi_0
hyperframe=6.1.0=pypi_0
idna=3.10=pypi_0
importlib-metadata=8.7.0=pypi_0
importlib-resources=6.5.2=pypi_0
instructor=1.10.0=pypi_0
ipython=8.37.0=pypi_0
itsdangerous=2.2.0=pypi_0
jedi=0.19.2=pypi_0
jinja2=3.1.6=pypi_0
jiter=0.10.0=pypi_0
joblib=1.5.2=pypi_0
json-repair=0.25.2=pypi_0
json5=0.12.1=pypi_0
jsonpatch=1.33=pypi_0
jsonpickle=4.1.1=pypi_0
jsonpointer=3.0.0=pypi_0
jsonref=1.1.0=pypi_0
jsonschema=4.25.0=pypi_0
jsonschema-specifications=2025.4.1=pypi_0
kiwisolver=1.4.9=pypi_0
kubernetes=33.1.0=pypi_0
lancedb=0.24.2=pypi_0
langchain=0.3.27=pypi_0
langchain-cohere=0.3.5=pypi_0
langchain-community=0.3.27=pypi_0
langchain-core=0.3.74=pypi_0
langchain-experimental=0.3.4=pypi_0
langchain-openai=0.2.14=pypi_0
langchain-text-splitters=0.3.9=pypi_0
langsmith=0.3.45=pypi_0
ld_impl_linux-64=2.44=h1423503_1
libexpat=2.7.1=hecca717_0
libffi=3.4.6=h2dba641_1
libgcc=15.1.0=h767d61c_4
libgcc-ng=15.1.0=h69a702a_4
libgomp=15.1.0=h767d61c_4
liblzma=5.8.1=hb9d3cd8_2
libnsl=2.0.1=hb9d3cd8_1
libsqlite=3.50.4=h0c1763c_0
libuuid=2.38.1=h0b41bf4_0
libxcrypt=4.4.36=hd590300_1
libzlib=1.3.1=hb9d3cd8_2
litellm=1.74.9=pypi_0
mako=1.3.10=pypi_0
markdown-it-py=4.0.0=pypi_0
markupsafe=3.0.2=pypi_0
marshmallow=3.26.1=pypi_0
matplotlib=3.10.5=pypi_0
matplotlib-inline=0.1.7=pypi_0
mdurl=0.1.2=pypi_0
mem0ai=0.1.116=pypi_0
mmh3=5.2.0=pypi_0
monotonic=1.6=pypi_0
mpmath=1.3.0=pypi_0
multidict=6.6.4=pypi_0
mypy-extensions=1.1.0=pypi_0
ncurses=6.5=h2d0b736_3
networkx=3.4.2=pypi_0
nodeenv=1.9.1=pypi_0
numpy=1.26.4=pypi_0
oauthlib=3.3.1=pypi_0
onnxruntime=1.22.0=pypi_0
openai=1.99.9=pypi_0
openpyxl=3.1.5=pypi_0
openssl=3.5.2=h26f9b46_0
opentelemetry-api=1.36.0=pypi_0
opentelemetry-exporter-otlp-proto-common=1.36.0=pypi_0
opentelemetry-exporter-otlp-proto-grpc=1.36.0=pypi_0
opentelemetry-exporter-otlp-proto-http=1.36.0=pypi_0
opentelemetry-instrumentation=0.57b0=pypi_0
opentelemetry-instrumentation-asgi=0.57b0=pypi_0
opentelemetry-instrumentation-fastapi=0.57b0=pypi_0
opentelemetry-proto=1.36.0=pypi_0
opentelemetry-sdk=1.36.0=pypi_0
opentelemetry-semantic-conventions=0.57b0=pypi_0
opentelemetry-util-http=0.57b0=pypi_0
orjson=3.11.2=pypi_0
overrides=7.7.0=pypi_0
packaging=25.0=pypi_0
pandas=2.3.1=pypi_0
parso=0.8.4=pypi_0
pdfminer-six=20250506=pypi_0
pdfplumber=0.11.7=pypi_0
pexpect=4.9.0=pypi_0
pillow=11.3.0=pypi_0
pip=25.2=pyh8b19718_0
portalocker=2.7.0=pypi_0
posthog=3.25.0=pypi_0
prompt-toolkit=3.0.51=pypi_0
propcache=0.3.2=pypi_0
protobuf=5.29.5=pypi_0
psutil=7.1.0=pypi_0
ptyprocess=0.7.0=pypi_0
pure-eval=0.2.3=pypi_0
pyarrow=21.0.0=pypi_0
pyasn1=0.6.1=pypi_0
pyasn1-modules=0.4.2=pypi_0
pybase64=1.4.2=pypi_0
pybind11=3.0.1=pypi_0
pycparser=2.22=pypi_0
pydantic=2.11.7=pypi_0
pydantic-core=2.33.2=pypi_0
pydantic-settings=2.10.1=pypi_0
pygments=2.19.2=pypi_0
pyjwt=2.10.1=pypi_0
pyparsing=3.2.3=pypi_0
pypdf=5.9.0=pypi_0
pypdf2=3.0.1=pypi_0
pypdfium2=4.30.0=pypi_0
pypika=0.48.9=pypi_0
pyproject-hooks=1.2.0=pypi_0
pyright=1.1.403=pypi_0
pysbd=0.3.4=pypi_0
python=3.10.18=hd6af730_0_cpython
python-dateutil=2.9.0.post0=pypi_0
python-dotenv=1.1.1=pypi_0
python-graphviz=0.21=pypi_0
pytube=15.0.0=pypi_0
pytz=2025.2=pypi_0
pyvis=0.3.2=pypi_0
pyyaml=6.0.2=pypi_0
qdrant-client=1.15.1=pypi_0
readline=8.2=h8c095d6_2
referencing=0.36.2=pypi_0
regex=2025.7.34=pypi_0
requests=2.32.4=pypi_0
requests-oauthlib=2.0.0=pypi_0
requests-toolbelt=1.0.0=pypi_0
rich=13.9.4=pypi_0
rpds-py=0.27.0=pypi_0
rsa=4.9.1=pypi_0
schema=0.7.7=pypi_0
scikit-learn=1.7.2=pypi_0
scipy=1.15.3=pypi_0
seaborn=0.13.2=pypi_0
setuptools=80.9.0=pyhff2d567_0
sgmllib3k=1.0.0=pypi_0
shellingham=1.5.4=pypi_0
six=1.17.0=pypi_0
sniffio=1.3.1=pypi_0
soupsieve=2.7=pypi_0
sqlalchemy=2.0.43=pypi_0
stack-data=0.6.3=pypi_0
starlette=0.47.2=pypi_0
structlog=25.4.0=pypi_0
sympy=1.14.0=pypi_0
tabulate=0.9.0=pypi_0
tenacity=9.1.2=pypi_0
threadpoolctl=3.6.0=pypi_0
tiktoken=0.11.0=pypi_0
tk=8.6.13=noxft_hd72426e_102
tokenizers=0.20.3=pypi_0
toml=0.10.2=pypi_0
tomli=2.2.1=pypi_0
tomli-w=1.2.0=pypi_0
torch=1.12.1+cu113=pypi_0
torchaudio=0.12.1+cu113=pypi_0
torchvision=0.13.1+cu113=pypi_0
tqdm=4.67.1=pypi_0
traitlets=5.14.3=pypi_0
typer=0.16.0=pypi_0
types-requests=2.32.4.20250809=pypi_0
typing-extensions=4.14.1=pypi_0
typing-inspect=0.9.0=pypi_0
typing-inspection=0.4.1=pypi_0
tzdata=2025.2=pypi_0
urllib3=2.5.0=pypi_0
uv=0.8.9=pypi_0
uvicorn=0.35.0=pypi_0
uvloop=0.21.0=pypi_0
watchfiles=1.1.0=pypi_0
wcwidth=0.2.13=pypi_0
websocket-client=1.8.0=pypi_0
websockets=15.0.1=pypi_0
werkzeug=3.1.3=pypi_0
wheel=0.45.1=pyhd8ed1ab_1
wrapt=1.17.3=pypi_0
yarl=1.20.1=pypi_0
zipp=3.23.0=pypi_0
zstandard=0.23.0=pypi_0
